requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
docker
pyyaml
boto3
requests
